# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      smartphone_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  models.ReviewRequest:
    properties:
      comment:
        type: string
      rating:
        type: integer
    type: object
...
  models.Smartphone:
    properties:
      description:
        type: string
      display_size:
        type: number
      id:
        type: integer
      image_path:
        type: string
      memory:
        type: integer
      model:
        type: string
      price:
        type: integer
      producer:
        type: string
      ram:
        type: integer
      ratings_count:
        type: integer
      ratings_sum:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API Server for Smartbuy application
  termsOfService: http://swagger.io/terms/
  title: Smartbuy API
  version: "1.0"
paths:
...
  /smartphones:
    get:
      consumes:
      - application/json
      description: Get a list of all smartphones or filter by IDs
      parameters:
      - description: Comma separated IDs (e.g. 1,2,3)
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Smartphone'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: List Smartphones
      tags:
      - smartphones
  /smartphones/{smartphone_id}:
    get:
      description: Get a specific smartphone by ID
      parameters:
      - description: Smartphone ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Smartphone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Get a Smartphone
      tags:
      - smartphones
  /smartphones/{smartphone_id}/reviews:
    get:
      description: Get all reviews for a specific smartphone
      parameters:
      - description: Smartphone ID
        in: path
        name: smartphone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: List Reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Smartphone ID
        in: path
        name: smartphone_id
        required: true
        type: integer
      - description: Review Body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Post a Review
      tags:
      - reviews
  /smartphones/{smartphone_id}/reviews/{review_id}:
    delete:
      description: Deletes a review of a certain smartphone
      parameters:
      - description: Smartphone ID
        in: path
        name: smartphone_id
        required: true
        type: integer
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletes a review
      tags:
      - reviews
    get:
      parameters:
      - description: Smartphone ID
        in: path
        name: smartphone_id
        required: true
        type: integer
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Get a Review
      tags:
      - reviews
    patch:
      consumes:
      - application/json
      parameters:
      - description: Smartphone ID
        in: path
        name: smartphone_id
        required: true
        type: integer
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      - description: Review Body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit a Review
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"